Move images utilities to a dedicated class
