+++ app/src/androidTest/java/ch/epfl/polybazaar/FillListingActivityTest.java
@@ -54,13 +54,13 @@ import static androidx.test.espresso.matcher.ViewMatchers.withId;
 import static androidx.test.espresso.matcher.ViewMatchers.withTagValue;
 import static androidx.test.espresso.matcher.ViewMatchers.withText;
 import static androidx.test.internal.runner.junit4.statement.UiThreadStatement.runOnUiThread;
-import static ch.epfl.polybazaar.Utilities.convertBitmapToString;
-import static ch.epfl.polybazaar.Utilities.convertBitmapToStringWithQuality;
-import static ch.epfl.polybazaar.Utilities.convertDrawableToBitmap;
-import static ch.epfl.polybazaar.Utilities.convertFileToString;
-import static ch.epfl.polybazaar.Utilities.convertStringToBitmap;
-import static ch.epfl.polybazaar.Utilities.resizeBitmap;
-import static ch.epfl.polybazaar.Utilities.resizeStringImageThumbnail;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.convertBitmapToString;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.convertBitmapToStringWithQuality;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.convertDrawableToBitmap;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.convertFileToString;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.convertStringToBitmap;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.resizeBitmap;
+import static ch.epfl.polybazaar.utilities.ImageUtilities.resizeStringImageThumbnail;
 import static ch.epfl.polybazaar.database.datastore.DataStoreFactory.useMockDataStore;
 import static ch.epfl.polybazaar.login.MockAuthenticator.TEST_USER_EMAIL;
 import static ch.epfl.polybazaar.login.MockAuthenticator.TEST_USER_PASSWORD;
